Refactor: RequiredArgsConstructor instead Autowired in Scheduler[박한솔]
