+++ basilium-server/src/main/java/basilium/basiliumserver/batchJobScheduler/ProductImageScheduler.java
@@ -2,6 +2,7 @@ package basilium.basiliumserver.batchJobScheduler;
 
 import basilium.basiliumserver.s3Storage.service.S3StorageService;
 import basilium.basiliumserver.service.product.ProductService;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.scheduling.annotation.Scheduled;
@@ -13,16 +14,19 @@ import java.util.stream.Collectors;
 
 //상품 이미지
 @Slf4j
+@RequiredArgsConstructor //오토와이어 제거
 @Component
 public class ProductImageScheduler {
     private final ProductService productService;
     private final S3StorageService s3StorageService;
-
+/*
     @Autowired
     public ProductImageScheduler(ProductService productService, S3StorageService s3StorageService) {
         this.productService = productService;
         this.s3StorageService = s3StorageService;
     }
+
+ */
 
 
     //@Scheduled(initialDelay = 60000, fixedDelay = 120000) // 1분 후에 시작하고, 그 이후에는 매 2분마다 실행
